Entrega Projeto Persistência e MVC [24E2_2]

+++ LiteStreaming.DataSeeders/Implementations/Administrative/DataSeederAdministrative.cs
@@ -3,8 +3,8 @@
 namespace DataSeeders.Administrative;
 public class DataSeederAdministrative : IDataSeeder
 {
-    private readonly RegisterContextAdministravtive _context;
-    public DataSeederAdministrative(RegisterContextAdministravtive context)
+    private readonly RegisterContextAdministrative _context;
+    public DataSeederAdministrative(RegisterContextAdministrative context)
     {
         _context = context;
     }
@@ -12,6 +12,12 @@ public class DataSeederAdministrative : IDataSeeder
     {
         try
         {
+            if (_context.Database.CanConnect())
+            {
+                Console.WriteLine("Banco de dados já existe. Não será recriado.");
+                return;
+            }
+
             _context.Database.EnsureDeleted();
             _context.Database.EnsureCreated();
